Refactor all os-related funcs
