+++ internal/vpn/errors.go
@@ -1,0 +1,7 @@
+package vpn
+
+import "errors"
+
+var (
+	errCouldFindDefaultNetworkGateway = errors.New("couldn't find default network gateway")
+)
