+++ internal/vpn/os_client_linux.go
@@ -1,0 +1,39 @@
+//+build linux
+package vpn
+
+import (
+	"bytes"
+	"fmt"
+	"net"
+	"os/exec"
+)
+
+const (
+	defaultNetworkGatewayCMD = "ip r | grep \"default via\" | awk '{print $3}'"
+)
+
+// DefaultNetworkGateway fetches system's default network gateway.
+func DefaultNetworkGateway() (net.IP, error) {
+	outBytes, err := exec.Command("sh", "-c", defaultNetworkGatewayCMD).Output() //nolint:gosec
+	if err != nil {
+		return nil, fmt.Errorf("error running command %s: %w", defaultNetworkGatewayCMD, err)
+	}
+
+	outBytes = bytes.TrimRight(outBytes, "\n")
+
+	outLines := bytes.Split(outBytes, []byte{'\n'})
+
+	for _, l := range outLines {
+		if bytes.Count(l, []byte{'.'}) != 3 {
+			// initially look for IPv4 address
+			continue
+		}
+
+		ip := net.ParseIP(string(l))
+		if ip != nil {
+			return ip, nil
+		}
+	}
+
+	return nil, errCouldFindDefaultNetworkGateway
+}
